Added email template, deploy logic and refactored backend code
Each email is now loaded from a template file and populated
properly. Email template path is stored in an environment variable
and email templates are copied to the server as a part of the static
files job. As changes to the template don't require a restart of the
backend. Models have been moved to a separate package